requirements.txt
streamlit
google-cloud-documentai
google-auth
google-api-core
